futable_backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
idna==3.10
lxml==5.3.0
Markdown==3.7
numpy==2.2.1
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
